Added lines and tool tips to the chart.
Tool tip displays the actual data from the current day when you hover over
the body of a candlestick.

Lines are broken, they do not display the correct prices or in the correct area.
However they have been added to this commit as they are still helpful when 
looking at the chart.