Refactored Account Class - Lots of changes
A portfolio is now a dictionary of (symbol, positionList) where positionList is a list
of Position objects.  The first Position object (positionList[0]) is always the most 
recent.

A Position allows the trader to buy more shares of a stock she currently owns 
or sell a partial amount of the stock she owns.   A Position keeps track of 
each trade and also maintains an average price-per-share.

A Position is opened by purchasing stock.  It is closed when all shares of that 
stock have been sold, including any additions of new stock.

This change does not implement stop or limit orders.  However, since it adds 
support for an extensive buy/sell history I chose to temporarily remove support
for stop/limit orders.